refactor: split tests into synchronous/ and asynchronous/ subdirectories
Mirror the mongoengine project structure: separate sync and async tests
into tests/synchronous/ and tests/asynchronous/. Shared fixtures stay
in tests/conftest.py; models.py and types.py stay at tests/ root.
Update all relative imports accordingly.
