+++ tests/synchronous/test_fields.py
@@ -1,6 +1,6 @@
 import pytest
 
-from . import nodes, nodes_async
+from . import nodes
 from graphene_mongo import AsyncMongoengineConnectionField
 from graphene_mongo.fields import MongoengineConnectionField
 
@@ -45,20 +45,3 @@ def test_field_args_with_unconverted_field():
 
     field_args = ["id", "name"]
     assert set(field.field_args.keys()) == set(field_args)
-
-
-@pytest.mark.asyncio
-async def test_default_resolver_with_colliding_objects_field_async():
-    field = AsyncMongoengineConnectionField(nodes_async.ErroneousModelAsyncNode)
-
-    connection = await field.default_resolver(None, {})
-    assert 0 == len(connection.iterable)
-
-
-@pytest.mark.asyncio
-async def test_default_resolver_connection_list_length_async(fixtures):
-    field = AsyncMongoengineConnectionField(nodes_async.ArticleAsyncNode)
-
-    connection = await field.default_resolver(None, {}, **{"first": 1})
-    assert hasattr(connection, "list_length")
-    assert connection.list_length == 1
